python_basico/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Deprecated==1.2.13
Flask==2.0.3
Flask-HTTPAuth==4.5.0
Flask-Limiter==2.2.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-resources==5.4.0
itsdangerous==2.1.1
Jinja2==3.0.3
jsonschema==4.4.0
limits==2.3.3
MarkupSafe==2.1.0
msg==0.7.8
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.7
pyrsistent==0.18.1
pytz==2021.3
redis==4.1.4
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0

yagmail~=0.15.277
keyring~=23.5.0